refactor(modal): rename the previous Modal to LegacyModal
