+++ packages/bezier-react/src/components/Modals/LegacyModal/Modal.stories.tsx
