+++ packages/bezier-react/src/components/Modals/LegacyModal/Modal.styled.ts
