+++ packages/bezier-react/src/components/Modals/LegacyModal/Modal.types.tsx
