+++ packages/bezier-react/src/components/Modals/index.ts
@@ -1,3 +1,3 @@
 export * from './BaseModal'
 export * from './ConfirmModal'
-export * from './Modal'
+export * from './LegacyModal'
